// react-components/src/types/enums.ts
export enum SortTypeStatus {
  empty = '',
  alive = 'alive',
  dead = 'dead',
  unknown = 'unknown',
}

export enum SortTypeGender {
  empty = '',
  female = 'female',
  male = 'male',
  genderless = 'genderless',
  unknown = 'unknown',
}
export enum SortTypeSpecies {
  Empty = '',
  Alien = 'Alien',
  Human = 'Human',
  Animal = 'Animal',
  Poopybutthole = 'Poopybutthole',
  Humanoid = 'Humanoid',
  Cronenberg = 'Cronenberg',
  unknown = 'unknown',
}

export enum ACTION_TYPE {
  SET_PAGE = 'SET_PAGE',
  SET_PAGES = 'SET_PAGES',
  SET_VALUE = 'SET_VALUE',
  SET_CHARACTERS = 'SET_CHARACTERS',
  SET_INFO = 'SET_INFO',
  SET_SORT_STATUS = 'SET_SORT_STATUS',
  SET_SORT_GENDER = 'SET_SORT_GENDER',
  SET_SORT_SPECIES = 'SET_SORT_SPECIES',
  SET_CARDS_FORM = 'SET_CARDS_FORM',
  SET_FORM_DATA = 'SET_FORM_DATA',
  SET_CHARACTER_DATA = 'SET_CHARACTER_DATA',
}